Hitmarker first implementation
Implemented hitarker UI feature (only on the assault rifle, or the Weaponshooting script). Need to find a better way to handle UI elements to avoid having to manually drag and drop reference on the other objects in the Unity editor. Right now if I wanted to spawn an other assault rifle at runtime it would be missing the UI references. UIHandler class to handle UI events?