Granade hitmarker + weapon damage bug
Added hitmarkers when hitting enemies with granades.

Found and (partially) solved a bug that made enemies take damage when passing over a weapon that was laying on the floor since the collision gets registered as if they were hit by a weapon thrown at them.
Now it's solved but it still happens when the weapons are laying on a slope. Added to backlog.