Grapple Hook 1st implementation
The grappling hook is activated by shooting a raycast after the correct input. If the raycast hits something in a certain range, the player will start to get pulled towards that point.
I tried to have the player momentum affect the trajectory, but for now the hook pulls the player to the hook with a little too much insistence. So the player is able to un-hook itself by pressing the jump button and kind of swing around like SpiderMan.
Would be interesting to be able to pull weapons toowards the player. But how to make this happen without spaghettying the code (aka write a specific script only used for weapons when they get hit by the hook)? Is it possible to describe a general behavior for objects when they are hit with a hook and calculate dinamically who is the puller and the pulled? Is it even a real problem?
Hook is instant and invisible for now :(

Fixed an issue that prevented player to take damage from own granades